refactor: generalize article-transitions to page-transitions for full-site ClientRouter
